Fix unusable firmware update screen
If an incompatible robot is connected from the firmware update settings screen, on the resulting screen the robot is not clickable. This is caused by the navigation: clicking the update button leaves the update fragment and enters an other update fragment, and it results in the following presenter registration and unregistration:
register(fragment1)
register(fragment2)
unregister(fragment1)
[user interacts with this state]

Fix this issue by making sure the presenter checks if it is unregistered by its owner.